+++ src/test/java/com/github/toolarium/icap/client/ICAPSocketTimeoutTest.java
@@ -11,7 +11,6 @@ import com.github.toolarium.icap.client.dto.ICAPMode;
 import com.github.toolarium.icap.client.dto.ICAPRequestInformation;
 import com.github.toolarium.icap.client.dto.ICAPResource;
 import java.io.ByteArrayInputStream;
-
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -41,7 +40,7 @@ public class ICAPSocketTimeoutTest {
             ByteArrayInputStream resourceInputStream = new ByteArrayInputStream(new byte[] {});
             ICAPClientFactory.getInstance().getICAPClient(LOCALHOST, port, SERVICENAME)
                  .validateResource(ICAPMode.RESPMOD, 
-                                   new ICAPRequestInformation("userb", "emptyfile"), 
+                                   new ICAPRequestInformation("userb", "emptyfile").addCustomHeader("Test", "Header"), 
                                    new ICAPResource("build/test-emptyfile.com", resourceInputStream, 0));
         } catch (Exception ioe) { // I/O error
             LOG.warn(RESOURCE_COULD_NOT_BE_ACCESSED + ioe.getMessage(), ioe);