+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ 1.3.6 ] - 2024-12-11
+## [ 1.3.6 ] - 2025-01-25
+### Added
+- Added support of custom headers. The reserved headers such as Host, Connection, User-Agent, Preview, Encapsulated, Allow can not be overwritten.
+
+### Fixed
+- Bugfix issue #15 to handle socket timeout properly.
 
 ## [ 1.3.5 ] - 2024-12-11
 ### Added